requirements.txt
python-rtmidi
mido
textual
pyyaml